I am a little uncomfortable with how the app loads. According to
http://stackoverflow.com/questions/3698200/window-onload-vs-document-ready

$(document).ready() happens before window.onload, which means it could
happens before lissa.figure.init().

Also, I think it just makes it easier to figure out what is going on
with a webapp when all the startup code happens in one place, in a file
called app.js. So I'm putting the startup code from controls.js and
lissa.js in $(document).ready() in app.js.